+++ context/app/default_config.py
@@ -1,8 +1,13 @@
+from datetime import timedelta
+
+
 class DefaultConfig(object):
     # This should be updated when app.conf is updated:
     # Test runs will see this config rather than app.conf.
     # Tests should not make API calls...
     # but they may expect certain keys to be present.
+
+    PERMANENT_SESSION_LIFETIME = timedelta(minutes=60)
 
     ENTITY_API_TIMEOUT = 5
     ENTITY_API_BASE = 'should-be-overridden'